added pro validations, renamed context file, deleted<router from app2.jsx>
